added swal

+++ src/Pages/MyVisa.jsx
@@ -2,6 +2,7 @@ import React, { useContext, useEffect, useState } from 'react';
 import { AuthContext } from '../Context/AuthProvider';
 import axios from 'axios';
 import { toast } from 'react-toastify';
+import Swal from 'sweetalert2';
 
 const MyVisa = () => {
     const {user} = useContext(AuthContext)
@@ -12,9 +13,27 @@ const MyVisa = () => {
     },[])
     console.log(user)
     const handleDelete = (id) => {
-        axios.delete(`http://localhost:3000/delete/${id}`)
-        .then(res => setMyVisa(myVisa.filter(visa => visa._id !== id)))
-        toast('Visa Deleted successfully')
+       
+        Swal.fire({
+          title: "Are you sure?",
+          text: "You won't be able to revert this!",
+          icon: "warning",
+          showCancelButton: true,
+          confirmButtonColor: "#3085d6",
+          cancelButtonColor: "#d33",
+          confirmButtonText: "Yes, delete it!"
+        }).then((result) => {
+          if (result.isConfirmed) {
+            Swal.fire({
+              title: "Deleted!",
+              text: "Your file has been deleted.",
+              icon: "success"
+              
+            });
+            axios.delete(`http://localhost:3000/delete/${id}`)
+            .then(res => setMyVisa(myVisa.filter(visa => visa._id !== id)))
+          }
+        });
         
     }
     return (
